Split command registry and implementation
This allows UI elements to iterate over a registry and display commands,
while the command implementations can reference those same UI elements
when building modal UI objects
