+++ tools/ctl/ide/commandreg/registry.go
@@ -1,7 +1,7 @@
 // Copyright 2025 Google LLC
 // SPDX-License-Identifier: Apache-2.0
 
-package commands
+package commandreg
 
 import (
 	"context"
